Split out `getFileExtension`
